requirements.txt
fastapi==0.95.0
asyncpg==0.27.0
alembic==1.10.2
sqlalchemy==2.0.8
pytest==7.2.2
gunicorn
sentry-sdk
httpx==0.23.3
psycopg2-binary==2.9.6
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asyncpg